09.10.16 23:11 new version kyrsach

+++ src/main/java/KyrsachAOD/ChoiceNumber.java
@@ -11,18 +11,18 @@ public class ChoiceNumber {;
      */
     public static int choice(int[] arrayNumbers, double[] arrayChances){
         Random random = new Random();
-        int previous=0;
+        double previous=0;
         int number = random.nextInt(1000);
-        if((number>=0)&&(number<(int)(arrayChances[0]*10))){
+        if((number>=0)&&(number<(arrayChances[0]*10))){
             return arrayNumbers[0];
         } else {
-            previous+=(int)(arrayChances[0]*10);
+            previous+=(arrayChances[0]*10);
         }
         for(int i=1;i<arrayChances.length;i++){
-            if((number>=previous)&&(number<(previous+(int)(arrayChances[i]*10)))){
+            if((number>=previous)&&(number<(previous+(arrayChances[i]*10)))){
                 return arrayNumbers[i];
             } else{
-                previous+=(int)(arrayChances[i]*10);
+                previous+=(arrayChances[i]*10);
             }
     }
     /**
